backend/src/modules/splicing_detection.py
import cv2
import numpy as np

def detect_splicing(image: np.ndarray):
    """
    Detects image splicing using a placeholder method.

    Args:
        image: The input image as a NumPy array.

    Returns:
        A tuple containing:
        - probability (float): The probability of splicing.
        - heatmap (np.ndarray): The heatmap of the spliced region.
    """
    # Placeholder: In a real implementation, we would use SIFT/ORB or a CNN.
    # For now, let's return a dummy probability and a simple gradient heatmap.
    probability = 0.75  # Dummy probability

    # Generate a simple gradient heatmap for demonstration purposes
    gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    sobelx = cv2.Sobel(gray_image, cv2.CV_64F, 1, 0, ksize=5)
    sobely = cv2.Sobel(gray_image, cv2.CV_64F, 0, 1, ksize=5)
    gradient_magnitude = np.sqrt(sobelx**2 + sobely**2)
    heatmap = cv2.normalize(gradient_magnitude, None, 0, 255, cv2.NORM_MINMAX).astype(np.uint8)

    return probability, heatmap
