backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
pillow
torch
tensorflow
scikit-image
matplotlib
pytest
requests
httpx
python-multipart
